requirements.txt
scapy~=2.4.4
events~=0.4
python-socketio~=4.6.0
aiohttp~=3.7.2
aiohttp-index~=0.1